rename the features file to the slices
