+++ src/slices/trendSlice.js
